+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # TBD
 
+# 0.14.1
+### Features
+* Add `NetworkContext.GetServices` to get a set of running service IDs
+* Add `NetworkContext.GetLambdas` to get a set of running Kurtosis Lambda IDs
+
 # 0.14.0
 ### Breaking Changes
 * Split method `NetworkContext.WaitForEndpointAvailability` to `NetworkContext.WaitForHttpGetEndpointAvailability` and `NetworkContext.WaitForHttpPostEndpointAvailability`